quantity now is decimal, map measurement unit

+++ colanta-backend/App/Orders/Infraestructure/Repositories/Dtos/SiesaOrderDto.cs
@@ -119,7 +119,7 @@
         public int C263IndObsequio {get; set;}
         public string C263UnidMedida {get; set;}
         [JsonConverter(typeof(Int32Converter))]
-        public int C263Cantidad {get; set;}
+        public decimal C263Cantidad {get; set;}
         public decimal C263Precio {get; set;}
         public string C263Notas {get; set;}
         public decimal C263Impuesto {get; set;}
